+++ akakoilib/string/manacher.cpp
@@ -1,8 +1,8 @@
 // r[i]: 中心 i の奇数長回文の最大半径(回文長 = 2*r[i]-1)
-vector<ll> Manacher(string s) {
+vector<int> Manacher(string s) {
   int n = s.size();
   int i = 0, j = 0;
-  vector<ll> r(n);
+  vector<int> r(n);
   while(i < n) {
     while(i >= j && i + j < n && s[i-j] == s[i+j]) j++;
     r[i] = j;
